src/main/webapp/app/entities/versao/versao-update.component.spec.ts
/* tslint:disable max-line-length */
import { vitest } from 'vitest';
import { shallowMount, type MountingOptions } from '@vue/test-utils';
import sinon, { type SinonStubbedInstance } from 'sinon';
import { type RouteLocation } from 'vue-router';

import dayjs from 'dayjs';
import VersaoUpdate from './versao-update.vue';
import VersaoService from './versao.service';
import { DATE_TIME_LONG_FORMAT } from '@/shared/composables/date-format';
import AlertService from '@/shared/alert/alert.service';

type VersaoUpdateComponentType = InstanceType<typeof VersaoUpdate>;

let route: Partial<RouteLocation>;
const routerGoMock = vitest.fn();

vitest.mock('vue-router', () => ({
  useRoute: () => route,
  useRouter: () => ({ go: routerGoMock }),
}));

const versaoSample = { id: 123 };

describe('Component Tests', () => {
  let mountOptions: MountingOptions<VersaoUpdateComponentType>['global'];
  let alertService: AlertService;

  describe('Versao Management Update Component', () => {
    let comp: VersaoUpdateComponentType;
    let versaoServiceStub: SinonStubbedInstance<VersaoService>;

    beforeEach(() => {
      route = {};
      versaoServiceStub = sinon.createStubInstance<VersaoService>(VersaoService);
      versaoServiceStub.retrieve.onFirstCall().resolves(Promise.resolve([]));

      alertService = new AlertService({
        i18n: { t: vitest.fn() } as any,
        bvToast: {
          toast: vitest.fn(),
        } as any,
      });

      mountOptions = {
        stubs: {
          'font-awesome-icon': true,
          'b-input-group': true,
          'b-input-group-prepend': true,
          'b-form-datepicker': true,
          'b-form-input': true,
        },
        provide: {
          alertService,
          versaoService: () => versaoServiceStub,
        },
      };
    });

    afterEach(() => {
      vitest.resetAllMocks();
    });

    describe('load', () => {
      beforeEach(() => {
        const wrapper = shallowMount(VersaoUpdate, { global: mountOptions });
        comp = wrapper.vm;
      });
      it('Should convert date from string', () => {
        // GIVEN
        const date = new Date('2019-10-15T11:42:02Z');

        // WHEN
        const convertedDate = comp.convertDateTimeFromServer(date);

        // THEN
        expect(convertedDate).toEqual(dayjs(date).format(DATE_TIME_LONG_FORMAT));
      });

      it('Should not convert date if date is not present', () => {
        expect(comp.convertDateTimeFromServer(null)).toBeNull();
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', async () => {
        // GIVEN
        const wrapper = shallowMount(VersaoUpdate, { global: mountOptions });
        comp = wrapper.vm;
        comp.versao = versaoSample;
        versaoServiceStub.update.resolves(versaoSample);

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(versaoServiceStub.update.calledWith(versaoSample)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', async () => {
        // GIVEN
        const entity = {};
        versaoServiceStub.create.resolves(entity);
        const wrapper = shallowMount(VersaoUpdate, { global: mountOptions });
        comp = wrapper.vm;
        comp.versao = entity;

        // WHEN
        comp.save();
        await comp.$nextTick();

        // THEN
        expect(versaoServiceStub.create.calledWith(entity)).toBeTruthy();
        expect(comp.isSaving).toEqual(false);
      });
    });

    describe('Before route enter', () => {
      it('Should retrieve data', async () => {
        // GIVEN
        versaoServiceStub.find.resolves(versaoSample);
        versaoServiceStub.retrieve.resolves([versaoSample]);

        // WHEN
        route = {
          params: {
            versaoId: '' + versaoSample.id,
          },
        };
        const wrapper = shallowMount(VersaoUpdate, { global: mountOptions });
        comp = wrapper.vm;
        await comp.$nextTick();

        // THEN
        expect(comp.versao).toMatchObject(versaoSample);
      });
    });

    describe('Previous state', () => {
      it('Should go previous state', async () => {
        versaoServiceStub.find.resolves(versaoSample);
        const wrapper = shallowMount(VersaoUpdate, { global: mountOptions });
        comp = wrapper.vm;
        await comp.$nextTick();

        comp.previousState();
        await comp.$nextTick();

        expect(routerGoMock).toHaveBeenCalledWith(-1);
      });
    });
  });
});
